So it begins...
I've already changed a couple things in this base build. For starters, I've revamped the ender-pearl effect to work well now. It will teleport Mario much greater distances whereas before it would fail and not move Mario. It's consistent now and I'm proud of it :D

I have fixed the Speed powerup to be usable. Instead of Mario helplessly flying everywhere when getting the boost, it's limited to just when he's walking and is actually really helpful as a buff to players now.