Fixes, Nuke, Gun additions
Re-added some missing hooks to fix main.lua functions. Added Nuke, added sniper HUD, and added FPS view scaling.

Main2.lua is pending removal. Unfortunately, despite how nice neat and optimized this is, it remains unfinished and I'm not up to the amount of work it'd take to finish it.